backend/requirements.txt
# Django and REST Framework
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<25.0
django-environ>=0.11,<1.0

# Microsoft 365 / Azure AD authentication
msal>=1.26,<2.0
requests>=2.31,<3.0

# Scheduling
APScheduler>=3.10,<4.0

# Database (optional - for PostgreSQL)
# psycopg2-binary>=2.9,<3.0

# Production server (optional)
# gunicorn>=21.0,<23.0

# Development/Testing
# pytest>=7.4,<9.0
# pytest-django>=4.7,<5.0
# black>=23.0,<25.0
# flake8>=6.1,<8.0
